Change microphone swipe behaviour
I tryed to change the microphone swipe behaviour but need more testing.
I thought it would be better to lock audio if swiping from left to right instead bottom top.
Problem is that when passing from "pressed" to "moving" mouse .x has a glich. Plus the horizontal image of the locker i made is not recognized

I also adjusted some animation and added a threshold for audio duration (500ms).